Add some tests for command line parsing
